pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "ADSMOD"
version = "2.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn[standard]==0.40.0",
    "torch==2.10.0+cu130",
    "keras==3.13.1", 
    "scikit-learn==1.8.0",    
    "numpy==2.4.1",
    "pandas==3.0.0",
    "tqdm==4.67.1",
    "SQLAlchemy==2.0.46",
    "psycopg[binary]==3.3.1",
    "python-dotenv==1.2.1",
    "httpx==0.27.2",    
    "python-multipart==0.0.12",
    "aiohttp==3.13.3",   
    "PubChemPy==1.0.5",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-playwright>=0.5",
    "psutil>=6.1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["adsmod"] 

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true 

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
# torchvision = { index = "pytorch-cu130" }
# torchaudio  = { index = "pytorch-cu130" }
